fixed some minor "warning: comparison between signed and unsigned integer expressions"

+++ rts/Rendering/UnitDrawer.cpp
@@ -124,7 +124,7 @@ CUnitDrawer::CUnitDrawer(): CEventClient("[CUnitDrawer]", 271828, false)
 	cloakAlpha3 = std::min(1.0f, cloakAlpha + 0.4f);
 
 	// load unit explosion generators
-	for (int unitDefID = 1; unitDefID < unitDefHandler->unitDefs.size(); unitDefID++) {
+	for (size_t unitDefID = 1; unitDefID < unitDefHandler->unitDefs.size(); unitDefID++) {
 		UnitDef* ud = unitDefHandler->unitDefs[unitDefID];
 
 		for (std::vector<std::string>::const_iterator it = ud->modelCEGTags.begin(); it != ud->modelCEGTags.end(); ++it) {
